server/requirements.txt
paho-mqtt==1.6.1
pymongo==4.0.1